add user confirmation before deleting recipe

+++ src/app/components/recipes/recipe-detail/recipe-detail.component.ts
@@ -40,7 +40,9 @@ export class RecipeDetailComponent implements OnInit {
   }
 
   deleteRecipe() {
-    this.recipeService.deleteRecipe(this.recipe);
-    this.router.navigate(['/recipes']);
+    if (confirm('Are you sure you want to delete this recipe?')) {
+      this.recipeService.deleteRecipe(this.recipe);
+      this.router.navigate(['/recipes']);
+    }
   }
 }
